requirements.txt
fastapi==0.115.6
psycopg2==2.9.10
SQLAlchemy==2.0.36
uvicorn==0.32.1
pydantic-settings==2.6.1
asyncpg==0.30.0
email_validator==2.2.0
